+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Extend interface mappings charset to support name prefixed with underscore
+- Introspection triggers are part of device triggers. Expose an API closer to other triggers.
 
 ## [1.0.3] - 2022-07-04